+++ lib/src/data/user_options/user_options_repository_impl.dart
@@ -87,7 +87,7 @@ class UserOptionsRepositoryImpl implements UserOptionsRepository {
                 UserOptionsArgs.knownGeneratedFiles.defaultValue),
             ..._parseGlob(arg.getStringList(UserOptionsArgs.addToKnownGeneratedFiles) ?? []),
           ],
-          useEmojis: arg.getBooleanOrDefault(UserOptionsArgs.useEmojis),
+          useEmojis: arg.getBooleanOrDefault(UserOptionsArgs.printEmojis),
         ),
       );
     } catch (e, s) {