+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 2.0.0
 
-- Output now has emojis by default. You can disable it by setting `use-emojis` to `false`.
+- Output now has emojis by default. You can disable it by setting `print-emojis` to `false`.
 - Output is now sorted: files missing tests will be shown at the end of the report
 - Change output style